let's keep is signed

+++ include/segment_tree.hpp
@@ -40,23 +40,23 @@ namespace nm
     class SegmentTree
     {
     private:
-        size_t n;
+        int32_t n;
         vector<T> tree;
         vector<T> auxiliary;
         U *integrator;
     
     protected:
-        bool propagate(size_t i);
-        void construct(std::vector<T> &data, size_t lo, size_t hi, size_t i);
-        T query_tree(size_t lo, size_t hi, size_t tlo, size_t thi, size_t i);
-        T update_tree(T value, size_t position, size_t lo, size_t hi, size_t i);
-        T update_tree(T value, size_t lo, size_t hi, size_t tlo, size_t thi, size_t i);
+        bool propagate(int32_t i);
+        void construct(std::vector<T> &data, int32_t lo, int32_t hi, int32_t i);
+        T query_tree(int32_t lo, int32_t hi, int32_t tlo, int32_t thi, int32_t i);
+        T update_tree(T value, int32_t position, int32_t lo, int32_t hi, int32_t i);
+        T update_tree(T value, int32_t lo, int32_t hi, int32_t tlo, int32_t thi, int32_t i);
     
     public:
         SegmentTree(std::vector<T> &data, U *integrator_struct = new U());
-        T query(size_t left, size_t right);
-        T update(T value, size_t position);
-        T update(T value, size_t left, size_t right);
+        T query(int32_t left, int32_t right);
+        T update(T value, int32_t position);
+        T update(T value, int32_t left, int32_t right);
         ~SegmentTree();
     };
 } // namespace nm
